refactor(utils): refactor `flattedChildren`
