+++ packages/components/descriptions/src/description.vue
@@ -27,6 +27,7 @@
 <script lang="ts" setup>
 // @ts-nocheck
 import { computed, provide, useSlots } from 'vue'
+import { flattedChildren } from '@element-plus/utils'
 import { useNamespace, useSize } from '@element-plus/hooks'
 import ElDescriptionsRow from './descriptions-row.vue'
 import { descriptionsKey } from './token'
@@ -47,19 +48,6 @@ const slots = useSlots()
 provide(descriptionsKey, props)
 
 const descriptionKls = computed(() => [ns.b(), ns.m(descriptionsSize.value)])
-
-const flattedChildren = (children) => {
-  const temp = Array.isArray(children) ? children : [children]
-  const res = []
-  temp.forEach((child) => {
-    if (Array.isArray(child.children)) {
-      res.push(...flattedChildren(child.children))
-    } else {
-      res.push(child)
-    }
-  })
-  return res
-}
 
 const filledNode = (node, span, count, isLast = false) => {
   if (!node.props) {
