+++ packages/utils/vue/vnode.ts
@@ -137,3 +137,16 @@ export const ensureOnlyChild = (children: VNodeArrayChildren | undefined) => {
   }
   return children[0]
 }
+
+export const flattedChildren = (children: VNodeNormalizedChildren) => {
+  const vNodes = Array.isArray(children) ? children : [children]
+  const result: any[] = []
+  vNodes.forEach((child: any) => {
+    if (Array.isArray(child.children)) {
+      result.push(...flattedChildren(child.children))
+    } else {
+      result.push(child)
+    }
+  })
+  return result
+}
